Auth-Broker/01-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
google-cloud-firestore==2.17.0
google-cloud-secret-manager==2.21.0
python-dotenv==1.0.0